refactor: split the conversion pipeline by stage
src/formats/index.ts held the whole pipeline: the specified-format
parse wrappers, auto-detection, aggregation dispatch, and Markdown
formatting dispatch. Each stage now has its own module, and index.ts
buffers the input and calls them in order. No behavior changes.
